\name{oil.idiss}
\alias{oil.idiss}
\docType{data}
\title{Interval-valued dissimilarity data based on Ichino's fats and oil data}
\description{This dataset is interval-valued dissimilarity data of 8 different fats and oil.
}
\usage{data(oil.idiss)}
\format{The interval-valued dissimilarity matrix (an object of class "array": \code{oil.idiss[1,,]}: the lower dissmilarity matrix; \code{oil.idiss[2,,]}: the upper dissmilarity matrix).    
}

\examples{
data(oil.idiss)

}
\keyword{datasets}